completed the purchase, new purchase and updated purchase with advanced features.

+++ src/pages/purchase/Purchase.jsx
@@ -6,7 +6,6 @@ import { DataGrid } from "@mui/x-data-grid";
 import { Link } from "react-router-dom";
 import { useNavigate } from "react-router-dom";
 import { useEffect, useState, useContext } from "react";
-
 import {
   collection,
   getDocs,
@@ -14,6 +13,8 @@ import {
   doc,
   deleteDoc,
   onSnapshot,
+  query,
+  where,
 } from "firebase/firestore";
 import { db } from "../../firebase";
 
@@ -22,23 +23,19 @@ const userColumns = [
   {
     field: "suppName",
     headerName: "Name",
-    width: 170,
+    width: 210,
   },
 
   {
     field: "suppPhone",
     headerName: "Phone",
-    width: 100,
+    width: 140,
   },
-  // {
-  //   field: "suppEmail",
-  //   headerName: "Email",
-  //   width: 150,
-  // },
+
   {
     field: "fuelType",
     headerName: "Fuel",
-    width: 70,
+    width: 80,
   },
   {
     field: "fuelTank",
@@ -53,7 +50,7 @@ const userColumns = [
   {
     field: "pricePerLitter",
     headerName: "Price/Litter",
-    width: 90,
+    width: 80,
   },
   {
     field: "totalPrice",
@@ -65,42 +62,81 @@ const userColumns = [
     headerName: "Date",
     width: 120,
   },
-  {
-    field: "status",
-    headerName: "Status",
-    width: 160,
-    renderCell: (params) => {
-      return (
-        <div className={`cellWithStatus ${params.row.status}`}>
-          {params.row.status}
-        </div>
-      );
-    },
-  },
 ];
 
 const Purchase = () => {
   const navigate = useNavigate();
+
+  return (
+    <div className="sales">
+      <Sidebar />
+      <div className="salesContainer">
+        <Navbar />
+        <div className="datatable">
+          <div className="datatableTitle ">
+            Purchase
+            <div
+              className="link"
+              onClick={() => {
+                navigate("/purchase/new-purchase");
+              }}
+            >
+              Add New
+            </div>
+          </div>
+        </div>
+        <div className="h-screen mx-[1rem]">
+          <PurchaseData buttons={true} />
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export const PurchaseData = ({ buttons }) => {
+  const navigate = useNavigate();
   const [data, setData] = useState([]);
+  const [purchaseStatus, setPurchaseStatus] = useState([]);
   const { purchaseId, SetPurchaseId } = useContext(PurchaseContext);
 
   useEffect(() => {
-    const fetchData = async () => {
-      const querySnapshot = await getDocs(collection(db, "purchase"));
-      const list = [];
-      querySnapshot.forEach((doc) => {
-        list.push({ id: doc.id, status: doc.data().status, ...doc.data() });
-      });
-      setData([]);
-
-      for (let i = 0; i < list.length; i++) {
-        setTimeout(() => {
-          setData((prevData) => [...prevData, list[i]]);
-        }, i * 1000);
+    const unsub = onSnapshot(
+      collection(db, "purchase"),
+      (snapShot) => {
+        let list = [];
+        snapShot.docs.forEach((doc) => {
+          list.push({ id: doc.id, ...doc.data() });
+        });
+        setData(list);
+      },
+      (error) => {
+        console.log(error);
       }
-    };
+    );
 
-    fetchData();
+    return () => {
+      unsub();
+    };
+  }, []);
+
+  //FETCH SELLS STATUS
+  useEffect(() => {
+    const unsub = onSnapshot(
+      collection(db, "purchase"),
+      (snapShot) => {
+        let list = [];
+        snapShot.docs.forEach((doc) => {
+          list.push(doc.data().status);
+        });
+        setPurchaseStatus(list);
+      },
+      (error) => {
+        console.log(error);
+      }
+    );
+    return () => {
+      unsub();
+    };
   }, []);
 
   const clickUser = (id) => {
@@ -113,11 +149,16 @@ const Purchase = () => {
   };
 
   const handleDelete = async (id) => {
-    try {
-      await deleteDoc(doc(db, "purchase", id));
-      setData(data.filter((item) => item.id !== id));
-    } catch (error) {
-      console.log(error);
+    const confirmed = confirm(
+      "Are you sure you want to delete this transection?"
+    );
+    if (confirmed) {
+      try {
+        await deleteDoc(doc(db, "purchase", id));
+        setData(data.filter((item) => item.id !== id));
+      } catch (error) {
+        console.log(error);
+      }
     }
   };
 
@@ -125,25 +166,25 @@ const Purchase = () => {
     {
       field: "action",
       headerName: "Action",
-      width: 130,
+      width: 140,
       renderCell: (params) => {
         return (
-          <div className="cellAction">
+          <div className="cellAction flex gap-2">
             {/* <Link to="" style={{ textDecoration: "none" }}>
-            <div className="viewButton" onClick={() => ""}>
-              View
-            </div>
-          </Link> */}
+              <div className="viewButton" onClick={() => ""}>
+                View
+              </div>
+            </Link> */}
             <Link to="/purchase/edit-purchase">
               <div
-                className="editButton"
+                className="py-[2px] px-3 rounded-md text-green-600 border-[1px] border-solid border-green-600"
                 onClick={() => editUserBtn(params.row.id)}
               >
                 Edit
               </div>
             </Link>
             <div
-              className="deleteButton"
+              className="py-[2px] px-3 rounded-md text-red-600 border-[1px] border-solid border-red-600 cursor-pointer"
               onClick={() => handleDelete(params.row.id)}
             >
               Delete
@@ -154,34 +195,61 @@ const Purchase = () => {
     },
   ];
 
-  return (
-    <div className="purchase">
-      <Sidebar />
-      <div className="purchaseContainer">
-        <Navbar />
-        <div className="datatable">
-          <div className="datatableTitle">
-            Purchase
+  const statusColumn = [
+    {
+      field: "status",
+      headerName: "Status",
+      width: 100,
+      renderCell: (params) => {
+        const index = data.findIndex((item) => item.id === params.row.id);
+        const status = purchaseStatus[index];
+
+        useEffect(() => {
+          const timeoutId = setTimeout(() => {
+            const nextIndex = index + 1;
+            if (nextIndex < data.length) {
+              setData((prevData) => {
+                const newData = [...prevData];
+                newData[nextIndex].status = purchaseStatus[nextIndex];
+                return newData;
+              });
+            }
+          }, (index + 1) * 1000);
+
+          return () => {
+            clearTimeout(timeoutId);
+          };
+        }, [index]);
+
+        return (
+          <div className="cellAction">
             <div
-              className="link"
-              onClick={() => {
-                navigate("/purchase/new-purchase");
-              }}
+              className={`status ${status} ${
+                status === "Pending"
+                  ? "bg-yellow-500/20 text-yellow-600 py-[.2rem] px-[.5rem] rounded-md"
+                  : "text-green-600 bg-green-500/20 py-[.2rem] px-[.5rem] rounded-md"
+              } `}
             >
-              Add New
+              {status}
             </div>
           </div>
-          <DataGrid
-            className="datagrid"
-            rows={data}
-            columns={userColumns.concat(actionColumn)}
-            pageSize={9}
-            rowsPerPageOptions={[9]}
-            // checkboxSelection
-          />
-        </div>
-      </div>
-    </div>
+        );
+      },
+    },
+  ];
+  let columns = userColumns.concat(statusColumn);
+  if (buttons) {
+    columns = columns.concat(actionColumn);
+  }
+
+  return (
+    <DataGrid
+      className="datagrid h-full overflow-x-hidden"
+      rows={data}
+      columns={columns}
+      pageSize={9}
+      rowsPerPageOptions={[9]}
+    />
   );
 };
